refactor(infrastructure): 도메인에 따라 패키지 분리
