+++ src/main/java/wooteco/subway/Infrastructure/line/LineDao.java
@@ -1,4 +1,4 @@
-package wooteco.subway.Infrastructure;
+package wooteco.subway.Infrastructure.line;
 
 import wooteco.subway.domain.Line;
