+++ src/main/java/wooteco/subway/application/SectionService.java
@@ -3,7 +3,7 @@ package wooteco.subway.application;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import wooteco.subway.Infrastructure.SectionDao;
+import wooteco.subway.Infrastructure.section.SectionDao;
 import wooteco.subway.domain.Section;
 import wooteco.subway.domain.Sections;
 import wooteco.subway.exception.constant.SectionNotRegisterException;
