+++ src/main/java/wooteco/subway/application/StationService.java
@@ -2,7 +2,7 @@ package wooteco.subway.application;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import wooteco.subway.Infrastructure.StationDao;
+import wooteco.subway.Infrastructure.station.StationDao;
 import wooteco.subway.domain.Station;
 import wooteco.subway.exception.constant.DuplicateException;
 import wooteco.subway.exception.constant.NotExistException;
