+++ src/test/java/wooteco/subway/Infrastructure/DbLineDaoTest.java
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
+import wooteco.subway.Infrastructure.line.DbLineDao;
 import wooteco.subway.domain.Line;
 
 import java.util.List;
