+++ src/test/java/wooteco/subway/Infrastructure/DbStationDaoTest.java
@@ -4,8 +4,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
+import wooteco.subway.Infrastructure.station.DbStationDao;
 import wooteco.subway.domain.Station;
 
 import java.util.List;
