+++ src/test/java/wooteco/subway/acceptance/LineAcceptanceTest.java
@@ -10,9 +10,9 @@ import org.junit.jupiter.params.provider.ValueSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.test.context.jdbc.Sql;
-import wooteco.subway.Infrastructure.DbLineDao;
-import wooteco.subway.Infrastructure.DbSectionDao;
-import wooteco.subway.Infrastructure.DbStationDao;
+import wooteco.subway.Infrastructure.line.DbLineDao;
+import wooteco.subway.Infrastructure.section.DbSectionDao;
+import wooteco.subway.Infrastructure.station.DbStationDao;
 import wooteco.subway.domain.Section;
 import wooteco.subway.domain.Station;
 import wooteco.subway.dto.request.LineRequest;
