+++ src/test/java/wooteco/subway/application/LineServiceTest.java
@@ -5,10 +5,10 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import wooteco.subway.Infrastructure.LineDao;
-import wooteco.subway.Infrastructure.MemoryLineDao;
-import wooteco.subway.Infrastructure.MemorySectionDao;
-import wooteco.subway.Infrastructure.SectionDao;
+import wooteco.subway.Infrastructure.line.LineDao;
+import wooteco.subway.Infrastructure.line.MemoryLineDao;
+import wooteco.subway.Infrastructure.section.MemorySectionDao;
+import wooteco.subway.Infrastructure.section.SectionDao;
 import wooteco.subway.domain.Line;
 import wooteco.subway.exception.constant.BlankArgumentException;
 import wooteco.subway.exception.constant.DuplicateException;
