+++ src/test/java/wooteco/subway/application/StationServiceTest.java
@@ -5,8 +5,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import wooteco.subway.Infrastructure.MemoryStationDao;
-import wooteco.subway.Infrastructure.StationDao;
+import wooteco.subway.Infrastructure.station.MemoryStationDao;
+import wooteco.subway.Infrastructure.station.StationDao;
 import wooteco.subway.domain.Station;
 import wooteco.subway.exception.constant.BlankArgumentException;
 import wooteco.subway.exception.constant.DuplicateException;
